refactor: move file mutations to dedicated file, improve typing
