+++ client/src/data-provider/Files/index.ts
@@ -1,1 +1,2 @@
 export * from './queries';
+export * from './mutations';
